move gitlab_tokens_actor() into a separate module
